Resolve Conflicts:
	console/src/components/Services/Data/Metadata/Metadata.js
	server/src-exec/Ops.hs
	server/src-lib/Hasura/GraphQL/Resolve.hs
	server/src-lib/Hasura/GraphQL/Resolve/BoolExp.hs
	server/src-lib/Hasura/GraphQL/Resolve/Select.hs
	server/src-lib/Hasura/GraphQL/Schema.hs
	server/src-lib/Hasura/RQL/DDL/Schema/Table.hs
	server/src-lib/Hasura/SQL/DML.hs
	server/src-lib/Hasura/SQL/Types.hs
	server/src-lib/Hasura/Server/Query.hs
	server/tests-py/test_graphql_queries.py

+++ console/src/components/Services/EventTrigger/Settings/Settings.js
@@ -20,8 +20,12 @@ class Settings extends Component {
     )[0];
     triggerSchema = triggerSchema ? triggerSchema : {};
 
-    const handleDeleteTrigger = () => {
-      dispatch(deleteTrigger(triggerName));
+    const handleDeleteTrigger = e => {
+      e.preventDefault();
+      const isOk = confirm('Are you sure?');
+      if (isOk) {
+        dispatch(deleteTrigger(triggerName));
+      }
     };
 
     return (
@@ -66,6 +70,25 @@ class Settings extends Component {
                     {triggerSchema.retry_interval > 1 ? 'seconds' : 'second'}
                   </td>
                 </tr>
+                {'headers' in triggerSchema ? (
+                  <tr>
+                    <td>Headers</td>
+                    <td>
+                      <AceEditor
+                        mode="json"
+                        theme="github"
+                        name="headers"
+                        value={JSON.stringify(triggerSchema.headers, null, 4)}
+                        minLines={4}
+                        maxLines={100}
+                        width="100%"
+                        showPrintMargin={false}
+                        showGutter={false}
+                        readOnly
+                      />
+                    </td>
+                  </tr>
+                ) : null}
                 <tr>
                   <td>Operation / Columns</td>
                   <td>
@@ -79,6 +102,7 @@ class Settings extends Component {
                       width="100%"
                       showPrintMargin={false}
                       showGutter={false}
+                      readOnly
                     />
                   </td>
                 </tr>
